CI: Bump Golang version to 1.11 (#7)
Update the version of Go used in CI from 1.6 (no longer builds many dependencies, breaking the build) to 1.11. Along the way, fix a `certlynx` format specifier bug caught by 1.11's improved linting. This fixes the build, :tada: 

It might be a good idea to enable cron auto-builds in the Travis UI for this repo to catch breakages sooner.