Basic autoreply greeter functionality for lopsbot. (#6)
This allows configuring an `-autoreply` message that is sent in-channel at most once per autoreplyPeriod (24 hrs) per nick in response to things said in channel. This can be used (for e.g.) to close an IRC channel by sending an informative message to anyone that tries to use the channel. 

Along the way I removed a stale binary I believe was checked in by mistake (a59e65a) and removed an unnecessary dependency on CFSSL for logging (b92280c)